fix: remove fragment and combine one monthly report activity
