+++ app/src/main/res/drawable/round_green_btn.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle" android:padding="10dp">
+    <solid android:color="@color/green"/>
+    <corners android:radius="10dp" />
+</shape>
