+++ app/src/main/res/drawable/round_yellow_view.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle" android:padding="10dp">
+    <solid android:color="@color/yellow_box"/>
+    <corners android:radius="25dp" />
+</shape>
